+++ backend_datalab/src/main/java/com/proyecto/datalab/repository/VariableRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.proyecto.datalab.entity.Variable;
 
 @Repository
-public interface VariableRepository extends JpaRepository<Variable, Long> {
+public interface VariableRepository extends JpaRepository<Variable, Integer> {
 
 }
